medical document is added

+++ database/migrations/2024_10_18_101557_create_service_lasers_table.php
@@ -18,7 +18,7 @@ class CreateServiceLasersTable extends Migration
             $table->string('title');
             $table->string('skin');
             $table->string('weight');
-            $table->unsignedTinyInteger('shot');
+            $table->unsignedBigInteger('shot');
             $table->softDeletes();
             $table->timestamps();
         });
